backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coverage==5.5
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
passlib==1.7.4
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
